pyproject.toml
[project]
name = "uci_online_retail"
version = "0.1.0"
description = "UCI online retail case for Vertica"
authors = [
    {name = "sao90", email = "s.ostervig@hotmail.com"},
]
dependencies = ["openpyxl>=3.1.5", "python-dotenv>=1.2.1", "pandas>=2.3.3", "matplotlib>=3.10.8", "pyarrow>=22.0.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[project.optional-dependencies]
test = [
    "pytest>=7.4.3",
    "pytest-cookies>=0.7.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "pytest-env>=1.1.3",
]


[tool.setuptools.packages]
find = {}

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$major.$minor.$patch$prerelease"
version_scheme = "pep440"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true
version_files = ["pyproject.toml:(?i)version",
                 "devops/configuration-release-dev.yaml:(?i)_IMAGE_VERSION",
                 "devops/configuration-release-test.yaml:(?i)_IMAGE_VERSION",
                 "devops/configuration-release-prod-experiments.yaml:(?i)_IMAGE_VERSION",
                 "devops/configuration-release-prod-shared.yaml:(?i)_IMAGE_VERSION",
                 "app_config/env_dev.yaml:(?i)container_tag",
                 "app_config/env_test.yaml:(?i)container_tag",
                 "app_config/env_prod_shared.yaml:(?i)container_tag",
                 "app_config/env_prod_experiments.yaml:(?i)container_tag"
                ]
bump_message = "chore(bump): $current_version → $new_version [skip-ci]"

# https://github.com/nat-n/poethepoet
[tool.poe]
include = "poe.toml"

[tool.pdm]
distribution = false


[dependency-groups]
dev = [
    "commitizen>=3.13.0",
    "poethepoet>=0.24.4",
    "ipykernel>=6.29.5",
    "pre-commit==4.0.1",
    "ruff>=0.11.13",
]
